feat(array.utils): modernize packaging — dual ESM/CJS exports, type:module
First package to adopt the modern packaging shape (per the Igalia
"Shipping Node.js packages in 2025" recommendations). Used as the
template the other published packages will follow next.

What changes
- `type: "module"` + an `exports` map with `types`/`import`/`require`/
  `default` conditions for `.`, `./defined`, `./onEmpty`, `./flatMap`.
  Legacy `./dist/*` wildcard is kept during the transition so any
  external consumer pinning to deep paths keeps working.
- Three-stage build (replaces the previous webpack-only pipeline):
    1. `tsc -p tsconfig.pkg.json --emitDeclarationOnly --declaration`
       — `.d.ts` files mirror the source layout
    2. `vite build` — ESM `.js` + CJS `.cjs` per entry. Rollup
       tree-shaking is *disabled* (`treeshake: false`) because the
       polyfill modules mutate `Array.prototype` and static analysis
       can't see that as a side-effect. The package is tiny; the cost
       is negligible. Downstream consumers still benefit from the
       package.json `sideEffects` allowlist for their own tree-shaking.
    3. `webpack --config webpack.config.cjs` — a single-entry UMD
       `dist/index.umd.js` for the script-tag consumer in
       `apps/web-js`. The index UMD already includes all polyfills via
       side-effects, so the previous separate `defined/index.js`
       script tag is redundant and removed.
- `webpack.config.js` renamed to `.cjs` because `type: "module"` would
  otherwise make Node parse it as ESM.

Consumer updates
- `apps/web-js/index.html`: `dist/index.js` → `dist/index.umd.js`,
  redundant `dist/defined/index.js` script removed.
- `apps/web-ts/src/{defined,onEmpty}.spec.ts`: switched from the
  internal-path imports (`'@hansogj/array.utils/dist/defined'`) to the
  exports-map paths (`'@hansogj/array.utils/defined'`).

Verified
- `pnpm run pre-commit` green: ws:ts, circularity, lint, 308/308 tests,
  31/31 harness tests
- web-cs (jest, webpack consumer) resolves array.utils through
  `exports.require` → `dist/index.cjs`
- web-ts (jest, webpack consumer) resolves through `exports.import` →
  `dist/index.js`
- web-js script-tag consumer serves `dist/index.umd.js` with all
  polyfills inlined
