+++ CLAUDE.md
@@ -60,7 +60,8 @@ pnpm-workspace monorepo (`pnpm-workspace.yaml` → `packages/*`) of independentl
 
 **Build pipelines differ by package — do not assume one shape:**
 
-- `array.utils`, `find-js`, `abonnement-js`, `maybe` — webpack + ts-loader (`build:wp`), UMD output with `libraryTarget: 'umd'`. The shared `webpack.build.js` at the root provides the rules; each package's `webpack.config.js` extends it with `entry`/`output`. Entry is `src/index.ts` → `dist/index.js` for the first three; `maybe` is the outlier — single-file `src/maybe.ts` → `dist/maybe.js` (filename preserved for the published 2.x contract and the script-tag consumer in `apps/web-js`).
+- `find-js`, `abonnement-js`, `maybe` — webpack + ts-loader (`build:wp`), UMD output with `libraryTarget: 'umd'`. The shared `webpack.build.js` at the root provides the rules; each package's `webpack.config.js` extends it with `entry`/`output`. Entry is `src/index.ts` → `dist/index.js` for the first two; `maybe` is the outlier — single-file `src/maybe.ts` → `dist/maybe.js` (filename preserved for the published 2.x contract and the script-tag consumer in `apps/web-js`).
+- `array.utils` — **modernized**: `type: "module"`, three-stage build (`tsc` for `.d.ts`, `vite build` for ESM `.js` + CJS `.cjs` per entry, `webpack` for a single UMD `index.umd.js` for the script-tag consumer in `apps/web-js`). `package.json` `exports` map exposes `.`, `./defined`, `./onEmpty`, `./flatMap` with `types`/`import`/`require`/`default` conditions; legacy `./dist/*` wildcard kept during the transition. `sideEffects` lists the polyfill-bearing entries so consumers can tree-shake the pure ones. Rollup tree-shaking is disabled in `vite.config.ts` (`treeshake: false`) — static analysis can't see that `defined/polyfill.ts` and `onEmpty/onEmpty.ts` mutate `Array.prototype`, and disabling is cheaper than maintaining `/*#__PURE__*/` annotations for a tiny package. This is the template the other published packages will follow as they're modernized.
 - `immer-reduxer` — `tsc` only (no webpack bundle), declarations emitted to `dist/`. `peerDependencies` cover `immer`, `redux`, `react-redux`.
 - `git-prompt` — TypeScript (CommonJS). Built with `@vercel/ncc` into three CLI binaries (`git-prompt-co`, `git-prompt-commit`, `git-prompt-retry`) exposed via `bin` in its `package.json`. ncc auto-discovers `tsconfig.json` in the package directory (not `tsconfig.pkg.json` like the other packages); `declaration: false` keeps the bundles from emitting stray `.d.ts` files. The `ts` script runs real `tsc --noEmit`.
