+++ apps/web-js/index.html
@@ -4,8 +4,7 @@
         <meta charset="utf-8" />
         <title>[TestUtil: Dinosaur JS]</title>
         <script src="@hansogj/find-js/dist/index.js"></script>
-        <script src="@hansogj/array.utils/dist/index.js"></script>
-        <script src="@hansogj/array.utils/dist/defined/index.js"></script>
+        <script src="@hansogj/array.utils/dist/index.umd.js"></script>
         <script src="@hansogj/maybe/dist/maybe.js"></script>
         <link href="/style.css" rel="stylesheet" />
     </head>
