+++ package.json
@@ -69,6 +69,7 @@
     "ts-loader": "9.6.0",
     "ts-node": "10.9.2",
     "typescript": "5.9.3",
+    "vite": "8.0.16",
     "webpack": "5.107.2",
     "webpack-cli": "7.0.3",
     "webpack-dev-server": "5.2.5"
